Manual test for ScheduledFireTimeUtc property of JobExecutionContext The test fails with the following output: (Test started at 09:35 UTC) (Computer put to sleep at 09:36 UTC) (Trigger should have fired at 09:37 UTC but could not because the computer was in sleep) (Computer woken up at 09:41 UTC) Event Scheduled for 03/06/2016 09:41:37 +00:00 Ticked at 03/06/2016 09:41:37 +00:00 with Difference of 00:00:00.0260000 Scheduled Fire Time Wrong - It Should Have Been Earlier Because of the Misfire   Expected: greater than 00:00:30   But was:  00:00:00.0260000
A fixed job was needed to implement this test, so a fixed job factory was also added, which can be used in any circumstance when only one job is executed for all triggers.
In order to ask the tester to carry out the manual part of this test at the right time a MessageBox is used, which is why the build script is modified a bit as well to include a Windows Forms reference.