Fix a race condition that could cause duplicate trigger firings.
If the same trigger was simultaneously picked up by SelectTriggerToAcquire() in two scheduler instances (A and B) and the firing of the trigger completes in instance A before UpdateTriggerStateFromOtherState() was called in instance B, then instance B will perform a duplicate firing.

This happens because instance A will have transitioned the trigger state to Acquired and then back to Waiting. The call to UpdateTriggerStateFromOtherState() in instance B will then set the trigger to Acquired a second time (because the state is Waiting) and the already-fired trigger will be returned.

Guard against this condition by checking that the next fire time hasn't changed when acquiring the trigger. In the example, instance B would fail to acquire the trigger because of the change made to the next fire time by instance A.