Register an ITypeLoadHelper by default in services.AddQuartz()
The SimpleTypeLoadHelper is uesd by default if no other ITypeLoadHelper is registered, but as the required DI type isn't registered, you get an InvalidOperationException on startup.

This registers the required service, as long as no other ITypeLoadHelper is added inside the call to AddQuartz().

Unfortunately we can't call UseSimpleTypeLoader(), as that would overwrite the configuration for any other configured type loader (while leaving the service registration intact)