Clarifying some wording, fixing a couple typos
There were a few unneeded commas, parentheses and a couple typos that were making this section a little difficult to read. 

I also wanted to add a little blurb about some of the symptoms that unhandled exceptions in jobs can lead to in order to make it easier for someone running into those kinds of issues to track down this specific section of documentation.