Removed Name, Group, JobName, JobGroup and FullName properties as they represent information that is already available through the Key and JobKey properties.
For triggers, those constructors that take a group or jobGroup as argument now throw an ArgumentNullException when either of these are null.

Update `.ctor(string name, string group)`of Key\<T> to throw an ArgumentNullException when null is passed for either argument. Update Name and Group properties of Key<T> to to throw an ArgumentNullException when attempting to set to null.

Fixes #1359.