Removed Name, Group, JobName, JobGroup and FullName properties from AbstractTrigger (#1363)
* Removed Name, Group, JobName, JobGroup and FullName properties as they represent information that is already available through the Key and JobKey properties.

For triggers, those constructors that take a group or jobGroup as argument now throw an ArgumentNullException when either of these are null.

Update `.ctor(string name, string group)`of Key\<T> to throw an ArgumentNullException when null is passed for either argument. Update Name and Group properties of Key<T> to to throw an ArgumentNullException when attempting to set to null.

Fixes #1359.

* Update changelog to reflect (breaking) changes.

* Reintroduce backing fields to make sure we remain backward compatible for serialization.

* Fix case where one of the triggers, but not both, is null (#1369)

Improve performance when comparing a trigger against itself, and when comparing next fire time.
Add corresponding benchmarks.

* Sync v4 with main (#1374)

* Do not use BOM for source files. (#1361)

* Protect us from breaking (binary) serialization for AbstractTrigger (#1373)

* Add tests to protect us from breaking (binary) serialization for AbstractTrigger.

* Fix complete test.

* Removed Name, Group, JobName, JobGroup and FullName properties as they represent information that is already available through the Key and JobKey properties.

For triggers, those constructors that take a group or jobGroup as argument now throw an ArgumentNullException when either of these are null.

Update `.ctor(string name, string group)`of Key\<T> to throw an ArgumentNullException when null is passed for either argument. Update Name and Group properties of Key<T> to to throw an ArgumentNullException when attempting to set to null.

Fixes #1359.

* Update changelog to reflect (breaking) changes.

* Reintroduce backing fields to make sure we remain backward compatible for serialization.

* Removed Name, Group, JobName, JobGroup and FullName properties as they represent information that is already available through the Key and JobKey properties.

For triggers, those constructors that take a group or jobGroup as argument now throw an ArgumentNullException when either of these are null.

Update `.ctor(string name, string group)`of Key\<T> to throw an ArgumentNullException when null is passed for either argument. Update Name and Group properties of Key<T> to to throw an ArgumentNullException when attempting to set to null.

Fixes #1359.

* Update changelog to reflect (breaking) changes.

* Reintroduce backing fields to make sure we remain backward compatible for serialization.

* Removed Name, Group, JobName, JobGroup and FullName properties as they represent information that is already available through the Key and JobKey properties.

For triggers, those constructors that take a group or jobGroup as argument now throw an ArgumentNullException when either of these are null.

Update `.ctor(string name, string group)`of Key\<T> to throw an ArgumentNullException when null is passed for either argument. Update Name and Group properties of Key<T> to to throw an ArgumentNullException when attempting to set to null.

Fixes #1359.

* Reintroduce backing fields to make sure we remain backward compatible for serialization.