Reduce need to construct a stored Type for all JobType operations and lean on serialized form more.
Enables maintenance tasks such as Job delete to operate when the Type can no longer be constructed.
JobTypeSerialized is stored and can be retrieved as Type still via a Lazy getter

Non-breaking change, but if code depended on exceptions being thrown for invalid job type retreival this may alter the code path.