Persist extended properties on trigger type changes.
UpdateTrigger would only update the extended trigger properties based on the type of the (new) provided trigger. If the type has changed from what was previously persisted, this will leave the trigger in an invalid state without the extended properties row.

If the type has changed, remove the extended properties for the old type and add extended properties for the new type.

This also includes a change to dispose the DbCommand used to update the trigger.