Refactor code to use TimeProvider abstraction (#2286)
Replace DateTime usage with TimeProvider.System
Replace implicit DateTime to DateTimeOffset conversions with explicit.
Add in Analyzer to ensure DateTime isn't used.
Remove a check for IsRUnningOnMono. As it is no longer one of the supported deployment frameworks.
Registers a TimeProvider in DI (but is not using it)
Refactor Datebuilder to take optional TimeProvider. Default to TimeProvider.System on null